fixed circular dependency caused by keeping the framework for the question in the same file as calling QuizPageSingle
