+++ components/question_components.js
@@ -1,5 +1,6 @@
 import axios from "axios";
 import React from "react";
+import { View, Text, Image, TouchableOpacity } from "react-native";
 
 
 // Load in a question, display it to console
@@ -20,4 +21,25 @@ export const QuestionLoader = async () => {
   } catch (error) {
     console.error(error);
   }
+};
+
+
+export const BasicQuestionDisplay = ({ questionIndex, question, onNextQuestion}) => {
+  
+  return (
+    <View>
+      <Text>Question {questionIndex + 1}</Text>
+      <Text>{question.question_text}</Text>
+
+      <Image source={{ uri: `data:image/png;base64,${question.image_encoding}`}} 
+      style={{ width: 200, height: 200 }}/>
+
+      {question.options.map((option, i) => (
+        <Text key={i}>{option}</Text>
+      ))}
+      <TouchableOpacity onPress={onNextQuestion}>
+        <Text>Next Question</Text>
+      </TouchableOpacity>
+    </View>
+  );
 };
